remove global href links from modal triggers
hitRoutes is only recorded when the router trigger true is used in the
App level link handler. All App.controller triggered rendering does not
register to hitRoutes and unsyncs with the currentView when closing
modals. Changed all modals to app.trigger rather than using href and
the App.link handler.